[refactoring] adapt testscriptfactory in scriptrunner
