+++ TestShell/TestShell.vcxproj
@@ -161,6 +161,7 @@
     </ClCompile>
     <ClCompile Include="shell_write.h" />
     <ClCompile Include="testscript.cpp" />
+    <ClCompile Include="testscriptfactory.cpp" />
     <ClCompile Include="testscript_test.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
@@ -169,6 +170,7 @@
     <ClInclude Include="File.h" />
     <ClInclude Include="iprocess_executor.h" />
     <ClInclude Include="process_executor.h" />
+    <ClInclude Include="testscriptfactory.h" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="shell.h" />
